fix: change all Telegram user ID columns from INTEGER to BIGINT
Root cause: Telegram assigns user IDs exceeding the 32-bit signed integer
range (e.g. 8457445258 > 2^31-1 = 2,147,483,647). Every table that stored
a Telegram user ID as PostgreSQL INTEGER caused
  psycopg.errors.NumericValueOutOfRange: integer out of range
on the /api/auth/link/confirm endpoint (via anti_abuse_events INSERT).

Changes:
- backend/app/models/: change Integer -> BigInteger for telegram_user_id /
  actor_telegram_user_id / owner_telegram_user_id / requester_telegram_user_id
  across anti_abuse_event, audit_event, auth_session, claim, item, user_profile
- backend/alembic/versions/20260409_0016_bigint_telegram_user_ids.py: new
  migration that ALTER COLUMNs all eight affected columns to BIGINT in the
  live PostgreSQL database
- bot/main.py: distinguish between 404 (code expired), 429 (rate-limited),
  5xx (server crash) and network errors in cmd_link; each returns a specific,
  accurate message to the user instead of the generic "may be expired" fallback
- frontend/src/pages/ProfilePage.tsx: add copyToClipboard() helper that tries
  navigator.clipboard.writeText() first and falls back to the legacy
  execCommand('copy') approach for HTTP pages / browsers without permission;
  copy button now shows an error notice when copy fails rather than silently
  doing nothing

+++ backend/app/models/audit_event.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, Index, Integer, JSON, String, func
+from sqlalchemy import BigInteger, DateTime, Index, Integer, JSON, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -18,7 +18,7 @@ class AuditEvent(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
     event_type: Mapped[str] = mapped_column(String(80), nullable=False, index=True)
-    actor_telegram_user_id: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
+    actor_telegram_user_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True, index=True)
     item_id: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
     claim_id: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
     details: Mapped[dict] = mapped_column(JSON, nullable=False, default=dict)
